add placeholder to other places
improve random broadcast and remove unnecessary scoreboard stuff
