+++ src/main/java/dev/projectg/geyserhub/module/scoreboard/ScoreboardManager.java
@@ -1,7 +1,7 @@
 package dev.projectg.geyserhub.module.scoreboard;
 
 import dev.projectg.geyserhub.GeyserHubMain;
-import dev.projectg.geyserhub.module.Placeholders;
+import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Objects;
 
 
-public class ScoreboardManager extends Placeholders {
+public class ScoreboardManager {
     public ScoreboardManager() {
     }
 
@@ -25,7 +25,8 @@ public class ScoreboardManager extends Placeholders {
 
     public static void createScoreboard(Player player) {
         Scoreboard board = Objects.requireNonNull(Bukkit.getServer().getScoreboardManager()).getNewScoreboard();
-        Objective objective = board.registerNewObjective("GeyserHub", "dummy", replaceValues(player, GeyserHubMain.getInstance().getConfig().getString("Scoreboard.Title")));
+        Objective objective = board.registerNewObjective("GeyserHub", "dummy", PlaceholderAPI.setPlaceholders(player, GeyserHubMain.getInstance().getConfig().getString("Scoreboard.Title", "GeyserHub")));
+
         objective.setDisplaySlot(DisplaySlot.SIDEBAR);
         List<String> text = GeyserHubMain.getInstance().getConfig().getStringList("Scoreboard.Lines");
 
@@ -33,7 +34,7 @@ public class ScoreboardManager extends Placeholders {
         int limit = Math.min(text.size(), 16);
 
         for (int index = 0; index < limit; index++) {
-            String formattedLine = replaceValues(player, text.get(index));
+            String formattedLine = PlaceholderAPI.setPlaceholders(player, text.get(index));
             Score score = objective.getScore(ChatColor.translateAlternateColorCodes('&', formattedLine));
             score.setScore(limit - index);
         }
